+++ src/app/src/screens/Dashboard/Model.ts
@@ -2,30 +2,56 @@
 import { Action, Thunk, action, thunk } from "easy-peasy";
 // project
 import { AppRegistry } from "../../domain/types";
+import { ContainerStateList } from "../../Types";
 
+export interface ContainerStats {
+  paused: number;
+  running: number;
+  exited: number;
+}
 export interface DashboardModelState {
-  containersCount: number;
+  containerStats: ContainerStats;
 }
 
 export interface DashboardModel extends DashboardModelState {
   // Actions
-  setContainersCount: Action<DashboardModel, number>;
+  setContainersStats: Action<DashboardModel, Partial<ContainerStats>>;
   // Thunks
-  containersFetchCount: Thunk<DashboardModel>;
+  containersFetchStats: Thunk<DashboardModel>;
 }
 
 export const createModel = (registry: AppRegistry): DashboardModel => ({
-  containersCount: 0,
+  containerStats: {
+    paused: 0,
+    running: 0,
+    exited: 0,
+  },
   // Actions
-  setContainersCount: action((state, value) => {
-    state.containersCount = value;
+  setContainersStats: action((state, value) => {
+    if (value.paused !== undefined) {
+      state.containerStats.paused = value.paused;
+    }
+    if (value.running !== undefined) {
+      state.containerStats.running = value.running;
+    }
+    if (value.exited !== undefined) {
+      state.containerStats.exited = value.exited;
+    }
   }),
 
   // Thunks
-  containersFetchCount: thunk(async (actions) =>
+  containersFetchStats: thunk(async (actions) =>
     registry.withPending(async () => {
+      // TODO: Optimize this - avoid loading all containers data, avoid multi-traversal
       const containers = await registry.api.getContainers();
-      actions.setContainersCount(containers.length);
+      const pausedContainers = containers.filter(c => c.DecodedState === ContainerStateList.PAUSED);
+      const runningContainers = containers.filter(c => c.DecodedState === ContainerStateList.RUNNING);
+      const exitedContainers = containers.filter(c => c.DecodedState === ContainerStateList.EXITED);
+      actions.setContainersStats({
+        paused: pausedContainers.length,
+        running: runningContainers.length,
+        exited: exitedContainers.length,
+      });
       return containers.length;
     })
   )